Add inviteflow to reinvite

+++ server/server_modules/reinvite/reinvite.post.js
@@ -2,8 +2,22 @@ const expect = require('@edgeguideab/expect');
 const userUtil = require('app/users/User');
 const logger = require('app/logger');
 const db = require('app/models');
+const uuidv1 = require('uuid/v1');
+const {
+  createDevice,
+  createPasswords,
+  mailAndLog
+} = require('app/invite/utilities');
 
 async function reinvite(req, res) {
+  let privateKey = req.session.privateKey;
+  let senderEmail = req.session.email;
+  let encryptedMasterPassword = req.session.mp;
+  let sender = {
+    Organization: { name: req.session.organization },
+    email: senderEmail
+  };
+
   const expectations = expect(
     {
       reinvite_email: 'string'
@@ -20,7 +34,12 @@ async function reinvite(req, res) {
       .end();
   }
   const email = req.body.reinvite_email;
+  if (sender.email === email) {
+    logger.log('silly', 'User tried to self-invite');
+    return res.status(500).send();
+  }
   let user;
+
   try {
     user = await userUtil.getUser(email);
     await clearTempkeys(user);
@@ -33,9 +52,29 @@ async function reinvite(req, res) {
       .end();
   }
 
-  //Do invite flow again
+  user.uuid = uuidv1();
+  await db.User.update({ uuid: user.uuid }, { where: { id: user.id } });
 
-  res.send('Hello World');
+  try {
+    let {
+      newEncryptedMasterPassword,
+      encryptedPrivateKey,
+      tempPassword
+    } = await createPasswords(privateKey, encryptedMasterPassword);
+
+    await createDevice(user, newEncryptedMasterPassword, encryptedPrivateKey);
+    await mailAndLog(user, sender, tempPassword);
+
+    return res
+      .send({
+        uuid: user.uuid,
+        tempPassword: tempPassword.toString('base64')
+      })
+      .end();
+  } catch (error) {
+    console.error(error);
+  }
+  res.status(500).send();
 }
 
 async function deleteDevices(user) {
@@ -48,11 +87,7 @@ async function deleteDevices(user) {
 
 async function clearTempkeys(user) {
   return await db.sequelize.query(
-    `
-    DELETE FROM TempKeys
-    where uuid 
-    IN (select Devices.deviceId from Devices where userid = :id)
-`,
+    'DELETE FROM TempKeys where uuid IN (select Devices.deviceId from Devices where userid = :id)',
     { replacements: { id: user.id } }
   );
 }
